Support all receiver signatures from the CloudEvents Go SDK
The HTTP and Event callback registration functions already take `interface{}`,
and the CloudEvents SDK already supports a much wider collection of signatures
(if this just let them through!).

With this change CloudEvent functions can reply, which enables invocations from
a Knative Broker to reply to create new events.

This adds a new test for replies.

Fixes: https://github.com/GoogleCloudPlatform/functions-framework-go/issues/58