fix: minimize dependencies on 3P libraries
This keeps the functions framework as lightweight as possible, and
avoids pulling in all of the Google Cloud client libraries even though
only one package is needed.

The new `internal/metadata` package is just a copy of
[google-cloud-go/functions/metadata](https://github.com/googleapis/google-cloud-go/tree/main/functions/metadata)

Minimizing what transitive dependencies the Functions Framework has is
important to make them work with buildpacks, where user function code
has to be bundled together with the functions framework. If a function's
dependency conflicts with a Functions Framework dependency and the
function is using vendor, then there can be undefined behavior at build
and runtime.