chore: fix prerun.sh to work with all Go versions
Use a template `go.mod` instead of running `go mod tidy` so that the
version of Go on the machine running `prerun.sh`  does not impact the `go.mod` that gets
generated.

When building the resulting function source code with `pack`, the
version of Go actually used to build will be depedent on the builder
image used.