fix: log CloudEvent function errors to stderr
CloudEvent functions currently return 500 when an error is returned, but
do not do anything with the error message. This logs the error message
to stderr so it is not lost.

Unlike Background Functions, the error message is NOT included in the
response body though.

Also, add fields for checking stderr to the unit tests.