feat: Added instructions for binding locally.
Added the ability to specify a hostname when executing locally.

The default example in the README now includes a flag to only bind
to the loopback device on a machine so firewall alerts shouldn't
trigger by default (and users won't expose functions to the LAN).

Existing use of the Start function including previous versions and
within the FF Buildpacks for GCF should continue to function.

See FF buildpack template for more details about how it invokes:
https://github.com/GoogleCloudPlatform/buildpacks/blob/main/cmd/go/functions_framework/template_v0.go